# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DIU YT Downloader API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface VideoFormat {
  formatId: string;
  quality: string;
  resolution: string;
  filesize?: number;
  ext: string;
  hasVideo: boolean;
  hasAudio: boolean;
}

export interface VideoInfo {
  id: string;
  title: string;
  thumbnail: string;
  channelName: string;
  duration: number;
  durationFormatted: string;
  viewCount?: number;
  formats: VideoFormat[];
}

export interface DownloadUrlResponse {
  downloadUrl: string;
  filename: string;
  quality: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type GetVideoInfoParams = {
  /**
   * YouTube video URL
   */
  url: string;
};

export type GetDownloadUrlParams = {
  /**
   * YouTube video URL
   */
  url: string;
  /**
   * Format ID for the selected resolution
   */
  formatId: string;
  /**
   * Quality label (e.g. 720p, 1080p)
   */
  quality: string;
};