# lib/api-zod/src/generated/types/downloadUrlResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DIU YT Downloader API
 * OpenAPI spec version: 0.1.0
 */

export interface DownloadUrlResponse {
  downloadUrl: string;
  filename: string;
  quality: string;
}