# lib/api-zod/src/generated/types/videoFormat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DIU YT Downloader API
 * OpenAPI spec version: 0.1.0
 */

export interface VideoFormat {
  formatId: string;
  quality: string;
  resolution: string;
  filesize?: number;
  ext: string;
  hasVideo: boolean;
  hasAudio: boolean;
}