# lib/api-zod/src/generated/types/videoInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DIU YT Downloader API
 * OpenAPI spec version: 0.1.0
 */
import type { VideoFormat } from "./videoFormat";

export interface VideoInfo {
  id: string;
  title: string;
  thumbnail: string;
  channelName: string;
  duration: number;
  durationFormatted: string;
  viewCount?: number;
  formats: VideoFormat[];
}